requirements.txt
django
djangorestframework
environs
pip freeze > requirements.txt
